updates ID's from Integers -> Longs

+++ services/src/main/java/dev/glabay/features/customer/CustomerController.java
@@ -30,7 +30,7 @@ public class CustomerController {
 
     @GetMapping("/{customerId}")
     private ResponseEntity<CustomerDto> getCustomerById(@PathVariable String customerId) {
-        var custId = Integer.parseInt(customerId);
+        var custId = Long.parseLong(customerId);
         return ResponseEntity.ok(customerService.getCustomerById(custId));
     }
